refactor: restructure jobs into subpackages
