+++ docs/glacium.utils.rst
@@ -49,8 +49,8 @@ can be saved to ``CSV`` using ``numpy.savetxt``.
 The higher level :func:`~glacium.utils.convergence.analysis` helper performs the
 aggregation, writes ``cl_cd_stats.csv`` and generates a plot of the coefficients.
 It is executed automatically by
-:class:`glacium.jobs.analysis_jobs.ConvergenceStatsJob` after
-:class:`~glacium.jobs.fensap_jobs.MultiShotRunJob` completes.  The repository
+:class:`glacium.jobs.analysis.ConvergenceStatsJob` after
+:class:`~glacium.jobs.fensap.MultiShotRunJob` completes.  The repository
 includes a small ``run_MULTISHOT`` directory with example data which can be
 analysed manually:
 
@@ -80,10 +80,10 @@ The :func:`~glacium.utils.convergence.analysis_file` helper performs the
 same operations for a single solver output file.  It is invoked automatically
 by the following job classes:
 
-* :class:`glacium.jobs.analysis_jobs.FensapConvergenceStatsJob`
-* :class:`glacium.jobs.analysis_jobs.Drop3dConvergenceStatsJob`
-* :class:`glacium.jobs.analysis_jobs.Ice3dConvergenceStatsJob`
-* :class:`glacium.jobs.analysis_jobs.AnalyzeMultishotJob`
+* :class:`glacium.jobs.analysis.FensapConvergenceStatsJob`
+* :class:`glacium.jobs.analysis.Drop3dConvergenceStatsJob`
+* :class:`glacium.jobs.analysis.Ice3dConvergenceStatsJob`
+* :class:`glacium.jobs.analysis.AnalyzeMultishotJob`
 
 You can run it manually with:
