+++ glacium/engines/__init__.py
@@ -4,7 +4,6 @@ from .engine_factory import EngineFactory
 from .base_engine import BaseEngine, XfoilEngine, DummyEngine
 from .pointwise import PointwiseEngine, PointwiseScriptJob
 from .fensap import FensapEngine, FensapScriptJob
-from .fluent2fensap import Fluent2FensapJob
 
 __all__ = [
     "BaseEngine",
@@ -14,7 +13,6 @@ __all__ = [
     "PointwiseScriptJob",
     "FensapEngine",
     "FensapScriptJob",
-    "Fluent2FensapJob",
     "EngineFactory",
 ]
