+++ glacium/jobs/analysis/__init__.py
@@ -1,0 +1,17 @@
+from .convergence_stats import ConvergenceStatsJob
+from .fensap_convergence_stats import FensapConvergenceStatsJob
+from .drop3d_convergence_stats import Drop3dConvergenceStatsJob
+from .ice3d_convergence_stats import Ice3dConvergenceStatsJob
+from .analyze_multishot import AnalyzeMultishotJob
+from .fensap_analysis import FensapAnalysisJob
+from .mesh_analysis import MeshAnalysisJob
+
+__all__ = [
+    "ConvergenceStatsJob",
+    "FensapConvergenceStatsJob",
+    "Drop3dConvergenceStatsJob",
+    "Ice3dConvergenceStatsJob",
+    "AnalyzeMultishotJob",
+    "FensapAnalysisJob",
+    "MeshAnalysisJob",
+]
