+++ glacium/jobs/fensap/__init__.py
@@ -1,0 +1,13 @@
+from .fensap_run import FensapRunJob
+from .drop3d_run import Drop3dRunJob
+from .ice3d_run import Ice3dRunJob
+from .multishot_run import MultiShotRunJob
+from .fluent2fensap import Fluent2FensapJob
+
+__all__ = [
+    "FensapRunJob",
+    "Drop3dRunJob",
+    "Ice3dRunJob",
+    "MultiShotRunJob",
+    "Fluent2FensapJob",
+]
