+++ glacium/jobs/pointwise/__init__.py
@@ -1,0 +1,7 @@
+from .gci import PointwiseGCIJob
+from .mesh2 import PointwiseMesh2Job
+
+__all__ = [
+    "PointwiseGCIJob",
+    "PointwiseMesh2Job",
+]
