+++ glacium/jobs/pointwise/gci.py
@@ -1,7 +1,8 @@
-"""Predefined Pointwise job classes."""
+from __future__ import annotations
 
 from pathlib import Path
 from glacium.engines.pointwise import PointwiseScriptJob
+
 
 class PointwiseGCIJob(PointwiseScriptJob):
     """Run the GCI grid script."""
@@ -9,12 +10,3 @@ class PointwiseGCIJob(PointwiseScriptJob):
     name = "POINTWISE_GCI"
     template = Path("POINTWISE.GCI.glf.j2")
     deps: tuple[str, ...] = ("XFOIL_THICKEN_TE",)
-
-class PointwiseMesh2Job(PointwiseScriptJob):
-    """Generate a second grid based on the GCI step."""
-
-    name = "POINTWISE_MESH2"
-    template = Path("POINTWISE.mesh2.glf.j2")
-    deps = ()
-
-
