+++ tests/test_fluent2fensap_default.py
@@ -2,7 +2,7 @@ import sys
 from pathlib import Path
 sys.path.insert(0, str(Path(__file__).resolve().parents[1]))
 
-from glacium.engines.fluent2fensap import Fluent2FensapJob
+from glacium.jobs.fensap import Fluent2FensapJob
 from glacium.engines.base_engine import BaseEngine
 from glacium.models.config import GlobalConfig
 from glacium.managers.path_manager import PathBuilder
